fix: nickname step 제거
