+++ apps/web/src/app/index/signup/page.tsx
@@ -3,7 +3,6 @@ import BirthDate from '@/components/signup/birthdate'
 import { ITEM_LIST } from '@/components/signup/constant'
 import Id from '@/components/signup/id'
 import Name from '@/components/signup/name'
-import NickName from '@/components/signup/nickname'
 import Password from '@/components/signup/password'
 import Phonenumber from '@/components/signup/phonenumber'
 import Sex from '@/components/signup/sex'
@@ -30,7 +29,6 @@ export default function SignupPage() {
     'sex',
     'birthDate',
     'name',
-    'nickname',
     'done',
   ] as const
 
@@ -128,12 +126,6 @@ export default function SignupPage() {
                 </Step>
                 <Step name="name">
                   <Name title="이름을 입력해 주세요" onNext={handleNext} />
-                </Step>
-                <Step name="nickname">
-                  <NickName
-                    title="닉네임을 입력해 주세요"
-                    onNext={handleNext}
-                  />
                 </Step>
                 <Step name="done">
                   <SignupDonePage />
