--- apps/web/src/components/signup/nickname.tsx
@@ -1,48 +1,0 @@
-import { SignupRequest, nicknameRegex } from '@/validation/zod'
-import { Button } from '@yaksok/ui/button'
-import { TextField } from '@yaksok/ui/text-field'
-import { useWatch } from 'react-hook-form'
-import SignupTitle from './signup-title'
-import { WithFormContext } from './type'
-import { withFormContext } from './with-form-context'
-
-function NickName({ methods, title, onNext }: WithFormContext) {
-  const { control, register, handleSubmit } = methods
-
-  // 욕설 비하 등 부적절한 단어 필터링 기능이 없음
-  const nicknameValue = useWatch({
-    control: control,
-    name: 'nickname',
-    defaultValue: '',
-  })
-
-  const isDisabled = nicknameRegex.test(nicknameValue)
-
-  const handleDone = handleSubmit((data: SignupRequest) => {
-    console.log('✅ SUBMIT CALLED', data)
-    onNext()
-  })
-
-  return (
-    <div>
-      <SignupTitle>{title}</SignupTitle>
-      <TextField
-        label="닉네임"
-        placeholder="2자 이상 6자 이하 입력해 주세요"
-        type="text"
-        message={{
-          regexError: '2자 이상 6자 이하로 입력해 주세요',
-        }}
-        regex={nicknameRegex}
-        {...register('nickname')}
-      />
-      <div className="mt-25">
-        <Button disabled={!isDisabled} onClick={handleDone}>
-          다음
-        </Button>
-      </div>
-    </div>
-  )
-}
-
-export default withFormContext(NickName)
